requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
whitenoise
python-dotenv==1.0.0
sqlparse==0.5.1
